requirements.txt
streamlit>=1.0
opencv-python
scikit-image
numpy
Pillow
